+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.9.1] - 2022-05-18
+### Added
+- gen_class.TimeFormat: Class to store times in pre-defined or user-defined time formats and retrieve with the same time setting. 
+- gen_class.ArgParser.delete_arg: Deletes a value from the args_array attribute.
+- gen_class.ArgParser.update_arg: Updates a value in the args_array attribute.
+- gen_class.ArgParser.insert_arg: Inserts a key and value into the args_array attribute.
+- gen_class.ArgParser.get_args_keys: Returns list of keys from the args_array attribute.
+- gen_class.ArgParser.get_args: Returns the args_array attribute.
+
+
 ## [2.9.1] - 2022-05-06
 ### Added
 - gen_class.ArgParser.get_val: Look for key in args_array and return value or default value.